Add checking, improve move safety checking
Add the checking behavior to the king, along with the behavior of being unable to move to pieces that would put it under attack. Checkmating still needs to be added however. Also fix bugs with validateSafe() in order to have proper outputs for the Pawn and to prevent a recursive loop for the King.